Lots of bounding box work.
When using an overlay geojson, don't include it in the bounding
box calculations to decide on the initial zoom level.
Do bounding box calculations per-file and return the bbox
from all the file read commands, though currently no one saves
or uses this information. Eventually, it should be smarter, and
use the overlay size if no other size is specified.

Also add bounding box dimensions to the argparsing test,
including a new test that reads the land ownership overlay.