More work on smarter argument parsing and bounding.
When specifying track file(s) plus a polygon overlay,
use the tracks but not the overlay for computing initial zoom.
But if only the overlay is specified, show all of it.
When specifying a known site pin location plus tracks,
zoom out to show all the tracks while still keeping the pin centered.
Plus more tests to try to keep this all working.