Fix drawing when starting from site selection dialog.
A recent GTK change has made drawing fail for the main window
if that window is created in a subprocess after the site selection
dialog was already shown in the parent process. So do the fork
before showing the site selection dialog, so both windows are
in the child process.