Combine waypoints at the same coordinate.
When adding a new waypoint, if one already exists with the exact
same coordinates, append the new waypoint name to the existing one,
rather than creating a new waypoint.
This helps when running pytopo on multiple image files since it's
not that uncommon to shoot several images from the same spot.