When clicking on a track, show info about the trackpoint.
If there's no nearby waypoint, it's useful to have some info about
the nearest trackpoint. For instance, you can use that to get the
time when that trackpoint was added, making it easier to calculate
how long a given segment of trail took.
(Eventually I hope pytopo will be able to show that info for track
segments, but that's a somewhat bigger job.)